订单模块维护

+++ src/main/java/com/soft1851/smart/campus/model/entity/SysOrder.java
@@ -26,7 +26,7 @@ public class SysOrder {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  Integer pkOrderId;
+    private Long pkOrderId;
 
     /**
      * 卡号
